FEAT(lazy): loading with adding moduels
Added auth and all the files needed for this folder. I have added modules and routing to each of my files in order to write a cleaner code.. however now I have some errors that I can not find out why... Going to view with fresh eyes in the moring.