Modules, Lazy loading, fav page
I finished adding and editing the modules in each of the files. I also implemented lazy loading and had to join office hours to help with all the errors. I moved the login page to link with the new Favorites page so one will be able to add new items they want to do onboard.